% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codalm_em.R
\name{coda_lm_ci}
\alias{coda_lm_ci}
\title{Bootstrap Confidence Intervals Linear Regression for Compositional Outcomes and Predictors}
\usage{
coda_lm_ci(
  yout,
  ypred,
  accelerate = FALSE,
  R = 500,
  ci_type = "perc",
  conf = 0.95,
  ...
)
}
\arguments{
\item{yout}{A matrix of compositional outcomes. Each row is an observation, and must sum to 1.
If any rows do not sum to 1, they will be renormalized}

\item{ypred}{A matrix of compositional predictors. Each row is an observation, and must sum to 1.
If any rows do not sum to 1, they will be renormalized}

\item{accelerate}{A logical variable, indicating whether or not to use the Squarem algorithm
for acceleration of the EM algorithm. Default is FALSE.}

\item{R}{The number of bootstrap reptitions to use. Default is 500}

\item{ci_type}{A character string with the type of bootstrap interval to be calculated. One of
"norm", "perc", "basic", or "bca". See the documentation for
\code{\link[boot]{boot.ci}} for more information. Default is "perc".}

\item{conf}{A scalar between 0 and 1 containing the confidence level of the required intervals.
Default is .95.}

\item{...}{Additional arguments to pass to the \code{\link[boot]{boot}}  function}
}
\value{
A list, with \code{ci_L} and
}
\description{
Implements the EM algorithm as described in Fiksel et al. (2020)
for transformation-free linear regression for compositional outcomes and predictors.
}
