% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independence_test.R
\name{codalm_indep_test}
\alias{codalm_indep_test}
\title{Permutation Test for Linear Independence Between Compositional Outcomes and Predictors}
\usage{
codalm_indep_test(
  yout,
  ypred,
  nperms = 500,
  init.seed = 123,
  accelerate = FALSE
)
}
\arguments{
\item{yout}{A matrix of compositional outcomes. Each row is an observation, and must sum to 1.
If any rows do not sum to 1, they will be renormalized}

\item{ypred}{A matrix of compositional predictors. Each row is an observation, and must sum to 1.
If any rows do not sum to 1, they will be renormalized}

\item{nperms}{The number of permutations. Default is 500.}

\item{init.seed}{The initial seed for the permutations. Default is 123.}

\item{accelerate}{A logical variable, indicating whether or not to use the Squarem algorithm for acceleration of the EM algorithm. Default is FALSE.}
}
\value{
The p-value for the indepence test
}
\description{
Implements the loss function based permutation test as described in Fiksel et al. (2020)
for a test of linear indepence between compositional outcomes and predictors.
}
